Util enums rework into external files under Utils.Enums namespace, references updated
