+++ RelhaxModpack/RelhaxModpack/Atlases/AtlasUtils.cs
@@ -1,4 +1,5 @@
 ﻿using RelhaxModpack.Utilities;
+using RelhaxModpack.Utilities.Enums;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
