+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/KnownFolder.cs
@@ -1,0 +1,55 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// Standard folders registered with the system. These folders are installed with Windows Vista
+    /// and later operating systems, and a computer will have only folders appropriate to it
+    /// installed.
+    /// </summary>
+    public enum KnownFolder
+        {
+            /// <summary>
+            /// The user's Contacts folder
+            /// </summary>
+            Contacts,
+            /// <summary>
+            /// The user's Desktop folder
+            /// </summary>
+            Desktop,
+            /// <summary>
+            /// The user's Documents folder
+            /// </summary>
+            Documents,
+            /// <summary>
+            /// The user's Downloads folder
+            /// </summary>
+            Downloads,
+            /// <summary>
+            /// The user's Favorites folder
+            /// </summary>
+            Favorites,
+            /// <summary>
+            /// The user's Links folder
+            /// </summary>
+            Links,
+            /// <summary>
+            /// The user's Music folder
+            /// </summary>
+            Music,
+            /// <summary>
+            /// The user's Pictures folder
+            /// </summary>
+            Pictures,
+            /// <summary>
+            /// The user's Pictures folder
+            /// </summary>
+            SavedGames,
+            /// <summary>
+            /// The user's SavedGames folder
+            /// </summary>
+            SavedSearches,
+            /// <summary>
+            /// The user's Videos folder
+            /// </summary>
+            Videos
+    }
+}
