+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/XmlLoadType.cs
@@ -1,0 +1,17 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// Xml document load type enumeration
+    /// </summary>
+    public enum XmlLoadType
+    {
+        /// <summary>
+        /// loading Xml from a file on disk
+        /// </summary>
+        FromFile,
+        /// <summary>
+        /// loading Xml from a text string
+        /// </summary>
+        FromString
+    }
+}
