+++ RelhaxModpack/RelhaxModpack/Utilities/MacroUtils.cs
@@ -3,45 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Utilities
 {
-    /// <summary>
-    /// The types of text macro replacements
-    /// </summary>
-    public enum ReplacementTypes
-    {
-        /// <summary>
-        /// Replacing macros with file paths
-        /// </summary>
-        FilePath,
-
-        /// <summary>
-        /// Replacing patch arguments of the patch object
-        /// </summary>
-        PatchArguementsReplace,
-
-        /// <summary>
-        /// Replacing modpack created macros (like [quote]) with the corresponding characters
-        /// </summary>
-        PatchFiles,
-
-        /// <summary>
-        /// Replacing literal interpretations of special characters like newline and tab with escaped versions
-        /// </summary>
-        TextEscape,
-
-        /// <summary>
-        /// Replacing escaped versions of special characters like newline and tab with the literal interpretations
-        /// </summary>
-        TextUnescape,
-
-        /// <summary>
-        /// Replacing zip path macros with absolute extraction paths
-        /// </summary>
-        ZipFilePath
-    }
-
     /// <summary>
     /// A utility class to handle macros
     /// </summary>
