+++ RelhaxModpack/RelhaxModpack/Windows/DatabaseUpdater.xaml.cs
@@ -15,6 +15,7 @@ using RelhaxModpack.Utilities;
 using RelhaxModpack.Xml;
 using RelhaxModpack.UI;
 using RelhaxModpack.Database;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Windows
 {
