+++ RelhaxModpack/RelhaxModpack/Windows/ModSelectionList.xaml.cs
@@ -18,6 +18,7 @@ using System.Text;
 using RelhaxModpack.Xml;
 using RelhaxModpack.Utilities;
 using RelhaxModpack.Database;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Windows
 {
